requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
openai
python-multipart
faiss-cpu
tiktoken
numpy